* some cleanup and modularization
