+++ public/classes/msg/Msg.php
@@ -1,18 +1,6 @@
 <?php
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- * Description of Msg
- *
- * @author ecv
- */
 class Msg {
-
     private static $instance;
     private $messages = [];
