+++ public/classes/view/Menu.php
@@ -1,16 +1,5 @@
 <?php
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- * Description of Menu
- *
- * @author Christian Voigt <chris at notjustfor.me>
- */
 class Menu {
     function getMenu($name = 'site') {
         $login = new Login();
