+++ public/modules/autoload.php
@@ -1,7 +1,8 @@
 <?php 
 require_once('login/Login.php');
-require_once('profile/Profile.php');
+require_once('login/Login_Password.php');
 require_once('login/Register.php');
+require_once('profile/Profile.php');
 require_once('settings/Settings.php');
 require_once('identity/Identity.php');
 require_once('activity/Activity.php');
