+++ public/modules/identity/Identity.php
@@ -1,32 +1,6 @@
 <?php
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- * Description of Identity
- *
- * @author Christian Voigt <chris at notjustfor.me>
- */
 class Identity {
-    function initEnv () {}
-    
-    function create_tables() {
-        $db = db::getInstance();
-        $sql = "CREATE TABLE identities (
-            id INT(6) UNSIGNED PRIMARY KEY AUTO_INCREMENT,
-            user_id INT(6),
-            identity_id INT(6),
-            name VARCHAR(64)
-        )";
-        $result = $db->query($sql);
-        echo $sql;
-    }
-    
-    function get() {}
     
     public function getCurrentIdentity() {
         return 0;
@@ -65,3 +39,4 @@ class Identity {
 }
 $identity = new Identity();
 $identity->initEnv();
+unset($identity);
